can save to the current level being edited (still very primitive

+++ GE2D/Camera.hpp
@@ -3,7 +3,7 @@
 
 class Camera {
 private:
-	int m_zoom;
+	float m_zoom;
 	Vector2D m_center = { 0, 0 };
 
 public:
